+++ workspaces/x2a/plugins/x2a/src/components/ModuleTable/Artifacts.tsx
@@ -30,15 +30,18 @@ const styles = makeStyles({
 export const ArtifactLink = ({
   artifact,
   targetRepoUrl,
+  targetRepoBranch,
 }: {
   artifact: Artifact;
+  // TODO: the targetRepoUrl is probably not needed, the artifact.value should contain full URL
   targetRepoUrl: string;
+  targetRepoBranch: string;
 }) => {
   const classes = styles();
   const { t } = useTranslation();
   return (
     <Link
-      to={buildArtifactUrl(artifact.value, targetRepoUrl)}
+      to={buildArtifactUrl(artifact.value, targetRepoUrl, targetRepoBranch)}
       target="_blank"
       rel="noopener noreferrer"
       key={artifact.id}
@@ -52,18 +55,24 @@ export const ArtifactLink = ({
 export const Artifacts = ({
   artifacts,
   targetRepoUrl,
+  targetRepoBranch,
 }: {
   artifacts: Artifact[];
   targetRepoUrl: string;
+  targetRepoBranch: string;
 }) => {
   // Keep it dense
   return (
     <div>
       {artifacts.map(artifact => (
-        <>
-          <ArtifactLink artifact={artifact} targetRepoUrl={targetRepoUrl} />
+        <span key={artifact.id}>
+          <ArtifactLink
+            artifact={artifact}
+            targetRepoUrl={targetRepoUrl}
+            targetRepoBranch={targetRepoBranch}
+          />
           <br />
-        </>
+        </span>
       ))}
     </div>
   );